+++ src/main/java/com/example/bobfriend/service/RecruitmentService.java
@@ -25,18 +25,6 @@ public class RecruitmentService {
     private final WritingReportRepository reportRepository;
     private final MemberService memberService;
 
-    @Transactional
-    public DetailResponse findById(Long recruitmentId) {
-        Recruitment recruitment = getRecruitment(recruitmentId);
-        return new DetailResponse(recruitment);
-    }
-
-    @Transactional
-    public Page<SimpleResponse> findAll(Pageable pageable) {
-        return recruitmentRepository.findAll(pageable)
-                .map(SimpleResponse::new);
-    }
-
 
     @Transactional
     public DetailResponse create(Create recruitmentRequestDto) {
@@ -60,63 +48,6 @@ public class RecruitmentService {
             throw new MemberNotAllowedException(currentMember.getEmail());
     }
 
-
-    @Transactional
-    public Page<SimpleResponse> findAllByRestaurantAddress(
-            String address, Pageable pageable) {
-        return recruitmentRepository
-                .findAllByAddress(address, pageable)
-                .map(SimpleResponse::new);
-    }
-
-
-    @Transactional
-    public Page<SimpleResponse> findAllAvailable(Pageable pageable) {
-        Member currentMember = memberService.getCurrentMember();
-        return recruitmentRepository
-                .findAllAvailable(currentMember, pageable)
-                .map(SimpleResponse::new);
-    }
-
-    @Transactional
-    public Addresses findAllLocations(Double latitude, Double longitude, Integer zoomLevel) {
-        Map<Address, Integer> addressMap = new HashMap();
-
-        // 0.05가 지도 상에서 대충 500m 정도
-        // 줌 레벨이 -2 ~ 12까지의 값을 가짐
-        double bound = 0.05 * (zoomLevel + 3);
-
-        List<Recruitment> allByLocation = recruitmentRepository.findAllByLocation(latitude, longitude, bound);
-        Iterator<Recruitment> iterator =
-                allByLocation.iterator();
-
-        while (iterator.hasNext()) {
-            Recruitment recruitment = iterator.next();
-            Address address =
-                    new Address(recruitment);
-            addressMap.put(address, addressMap.getOrDefault(address, 0) + 1);
-        }
-
-        for (Map.Entry<Address, Integer> entry :
-                addressMap.entrySet()) {
-            entry.getKey().setCount(entry.getValue());
-        }
-        return new Addresses(new ArrayList<>(addressMap.keySet()));
-    }
-
-    @Transactional
-    public Page<SimpleResponse> findAllJoined(Pageable pageable) {
-        Member author = memberService.getCurrentMember();
-        return recruitmentRepository.findAllJoined(author, pageable)
-                .map(SimpleResponse::new);
-    }
-
-    @Transactional
-    public Page<SimpleResponse> findMyRecruitments(Pageable pageable) {
-        Member author = memberService.getCurrentMember();
-        return recruitmentRepository.findAllByAuthor(author, pageable)
-                .map(SimpleResponse::new);
-    }
 
     @Transactional
     public DetailResponse close(Long recruitmentId) {
@@ -175,6 +106,7 @@ public class RecruitmentService {
                 .map(DetailResponse::new);
     }
 
+
     public Page<DetailResponse> searchRestaurant(Condition.Search search, Pageable pageable) {
         return recruitmentRepository
                 .searchByRestaurant(search, pageable)
@@ -194,6 +126,7 @@ public class RecruitmentService {
         Member member = memberService.getCurrentMember();
         reportService.reportWriting(member, recruitment);
     }
+
 
     @Transactional
     public void expire() {
@@ -216,6 +149,7 @@ public class RecruitmentService {
         return recruitment;
     }
 
+
     private boolean checkSexRestriction(Recruitment recruitment,
                                         Member member) {
         Sex restriction = recruitment.getSexRestriction();
@@ -225,6 +159,7 @@ public class RecruitmentService {
         }
         return false;
     }
+
 
     private Recruitment checkAppointmentTime(Recruitment recruitment) {
         if (LocalDateTime.now().isAfter(recruitment.getEndAt()))
