+++ src/test/java/com/example/bobfriend/service/RecruitmentServiceTest.java
@@ -100,64 +100,19 @@ class RecruitmentServiceTest {
     }
 
 
-    @Test
-    public void findByIdSuccess() {
-        given(recruitmentRepository.findById(testRecruitment.getId()))
-                .willReturn(Optional.ofNullable(testRecruitment));
-
-        DetailResponse byId = recruitmentService
-                .findById(testRecruitment.getId());
-
-        DetailResponse dtoFromEntity =
-                new DetailResponse(testRecruitment);
-
-        assertThat(byId, equalTo(dtoFromEntity));
-    }
-
-
-    @Test
-    public void findByIdFail() {
-        given(recruitmentRepository.findById(0L))
-                .willReturn(Optional.empty());
-
-        assertThrows(RecruitmentNotFoundException.class,
-                () -> recruitmentService.findById(0L)
-        );
-
-    }
-
-
-    @Test
-    public void findAll() {
-        List<Recruitment> recruitmentList = Arrays.asList(testRecruitment);
-        PageRequest pageRequest = PageRequest.of(0, 1);
-        List<SimpleResponse> collect = recruitmentList.stream()
-                .map(r -> new SimpleResponse(r))
-                .collect(Collectors.toList());
-        Page<SimpleResponse> page = new PageImpl<>(collect);
-
-        given(recruitmentRepository.findAll(pageRequest))
-                .willReturn(new PageImpl<>(recruitmentList));
-        Page<SimpleResponse> responseDtoList = recruitmentService.findAll(pageRequest);
-
-        assertThat(responseDtoList,
-                equalTo(page));
-    }
-
 
     @Test
     public void create() {
         when(memberService.getCurrentMember()).thenReturn(testAuthor);
         when(recruitmentRepository.save(any()))
                 .thenReturn(testRecruitment);
-        when(recruitmentRepository.findById(any())).thenReturn(Optional.ofNullable(testRecruitment));
+
         Create requestDto = new Create(testRecruitment);
 
         DetailResponse add = recruitmentService.create(requestDto);
 
-        DetailResponse byId = recruitmentService.findById(testRecruitment.getId());
-
-        assertThat(add, equalTo(byId));
+        assertThat(add.getAuthor().getId(),
+                equalTo( testRecruitment.getAuthor().getId()));
     }
 
 
@@ -213,163 +168,6 @@ class RecruitmentServiceTest {
     }
 
 
-    @Test
-    public void findAllJoinedRecruitments() {
-        Member testMember = Member.builder()
-                .id(1)
-                .email("testMember@test.com")
-                .nickname("testMember")
-                .password("testPassword")
-                .sex(Sex.FEMALE)
-                .birth(LocalDate.now())
-                .active(true)
-                .rating(0.0)
-                .numberOfJoin(0)
-                .build();
-
-        Recruitment recruitment = Recruitment.builder()
-                .id(1000L)
-                .title("addedRecruitment")
-                .content("")
-                .author(testAuthor)
-                .members(new HashSet<>(Arrays.asList(testMember)))
-                .comments(List.of(testComment))
-                .totalNumberOfPeople(3)
-                .restaurantName("testRestaurantName")
-                .restaurantAddress("testRestaurantAddress")
-                .latitude(0.0)
-                .longitude(0.0)
-                .createdAt(LocalDateTime.now())
-                .appointmentTime(LocalDateTime.now().plusHours(4))
-                .endAt(LocalDateTime.now().plusDays(1))
-                .active(true)
-                .build();
-
-        when(memberService.getCurrentMember()).thenReturn(testMember);
-        PageRequest pageRequest = PageRequest.of(0, 1);
-
-        List<Recruitment> recruitmentList = Arrays.asList(testRecruitment, recruitment);
-        given(recruitmentRepository.findAllJoined(any(), any()))
-                .willReturn(new PageImpl<>(recruitmentList));
-
-        Page<SimpleResponse> responseDtoList =
-                recruitmentService.findAllJoined(pageRequest);
-
-        assertThat(responseDtoList.toList(),
-                equalTo(recruitmentList.stream()
-                        .map(r -> new SimpleResponse(r))
-                        .collect(Collectors.toList())));
-    }
-
-
-    @Test
-    public void findAllAvailableRecruitments() {
-        Member testMember = Member.builder()
-                .id(1)
-                .email("testMember@test.com")
-                .nickname("testMember")
-                .password("testPassword")
-                .sex(Sex.FEMALE)
-                .birth(LocalDate.now())
-                .active(true)
-                .rating(0.0)
-                .numberOfJoin(0)
-                .build();
-        Recruitment recruitment = Recruitment.builder()
-                .id(1000L)
-                .title("addedRecruitment")
-                .content("")
-                .author(testAuthor)
-                .members(new HashSet<>())
-                .comments(new LinkedList<>())
-                .totalNumberOfPeople(3)
-                .restaurantName("testRestaurantName")
-                .restaurantAddress("testRestaurantAddress")
-                .latitude(0.0)
-                .longitude(0.0)
-                .createdAt(LocalDateTime.now())
-                .appointmentTime(LocalDateTime.now().plusHours(4))
-                .endAt(LocalDateTime.now().plusDays(1))
-                .active(true)
-                .build();
-        when(memberService.getCurrentMember()).thenReturn(testMember);
-        PageRequest pageRequest = PageRequest.of(0, 1);
-        List<Recruitment> recruitmentList = Arrays.asList(recruitment);
-        given(recruitmentRepository.findAllAvailable(any(), any()))
-                .willReturn(new PageImpl<>(recruitmentList));
-        Page<SimpleResponse> responseDtoList =
-                recruitmentService.findAllAvailable(pageRequest);
-
-        assertThat(responseDtoList.toList(),
-                equalTo(Arrays.asList(new SimpleResponse(recruitment))));
-    }
-
-
-    @Test
-    public void findMyRecruitments() {
-        Member testMember = Member.builder()
-                .id(1)
-                .email("testMember@test.com")
-                .nickname("testMember")
-                .password("testPassword")
-                .sex(Sex.FEMALE)
-                .birth(LocalDate.now())
-                .active(true)
-                .rating(0.0)
-                .numberOfJoin(0)
-                .build();
-        Recruitment recruitment = Recruitment.builder()
-                .id(1000L)
-                .title("addedRecruitment")
-                .content("")
-                .author(testAuthor)
-                .members(new HashSet<>(Arrays.asList(testMember)))
-                .comments(List.of(testComment))
-                .totalNumberOfPeople(3)
-                .restaurantName("testRestaurantName")
-                .restaurantAddress("testRestaurantAddress")
-                .latitude(0.0)
-                .longitude(0.0)
-                .createdAt(LocalDateTime.now())
-                .appointmentTime(LocalDateTime.now().plusHours(4))
-                .endAt(LocalDateTime.now().plusDays(1))
-                .active(true)
-                .build();
-        when(memberService.getCurrentMember()).thenReturn(testMember);
-
-        PageRequest pageRequest = PageRequest.of(0, 1);
-
-        List<Recruitment> recruitmentList = Arrays.asList(testRecruitment, recruitment);
-        given(recruitmentRepository.findAllByAuthor(any(), any()))
-                .willReturn(new PageImpl<>(recruitmentList));
-
-        Page<SimpleResponse> responseDtoList =
-                recruitmentService.findMyRecruitments(pageRequest);
-
-        assertThat(responseDtoList.toList(),
-                equalTo(recruitmentList.stream()
-                        .map(r -> new SimpleResponse(r))
-                        .collect(Collectors.toList())));
-    }
-
-
-    @Test
-    void findAllByRestaurantAddress() {
-        PageRequest pageRequest = PageRequest.of(0, 1);
-        when(recruitmentRepository.findAllByAddress(
-                any(), any()
-        )).thenReturn(new PageImpl<>(Arrays.asList(testRecruitment)));
-
-        String restaurantAddress = "restaurantAddress";
-
-        Page<SimpleResponse> restaurantList = recruitmentService
-                .findAllByRestaurantAddress(restaurantAddress, pageRequest);
-
-        assertThat(restaurantList.toList(),
-                equalTo(Arrays.asList(
-                        new SimpleResponse(testRecruitment)
-                )));
-    }
 
 
     @Test
@@ -436,33 +234,6 @@ class RecruitmentServiceTest {
     }
 
 
-    @Test
-    void findAllLocationsTest() {
-        double lat = 33.4566084914484;
-        double lon = 126.56207301534569;
-
-        List<Recruitment> recruitments = new ArrayList<>();
-        for (int i = 0; i < 30; i++) {
-            Recruitment recruitment = Recruitment.builder()
-                    .latitude(lat)
-                    .longitude(lon)
-                    .restaurantAddress("test")
-                    .author(testAuthor)
-                    .appointmentTime(LocalDateTime.now())
-                    .sexRestriction(Sex.NONE)
-                    .active(true)
-                    .build();
-            recruitments.add(recruitment);
-        }
-        when(recruitmentRepository.findAllByLocation(any(), any(), any()))
-                .thenReturn(recruitments);
-
-        Address address = new Address(recruitments.get(0));
-        address.setCount(recruitments.size());
-        Addresses allLocations = recruitmentService.findAllLocations(lat, lon, 3);
-
-        assertThat(allLocations.getAddresses(), equalTo(List.of(address)));
-    }
 
 
 }
